refactor: altera o nome do package e do artifactId
o nome do package agora é edu.grupo8 e o do artifactId é keepfix
