+++ src/main/java/edu/grupo8/MainLayoutController.java
@@ -1,4 +1,4 @@
-package com.ricardo;
+package edu.grupo8;
 
 import java.io.IOException;
 import java.net.URL;
@@ -6,8 +6,7 @@ import java.util.ResourceBundle;
 
 import org.kordamp.ikonli.javafx.Icon;
 
-import com.ricardo.components.MenuHighlightShape;
-
+import edu.grupo8.components.MenuHighlightShape;
 import javafx.animation.FadeTransition;
 import javafx.animation.Interpolator;
 import javafx.animation.KeyFrame;
@@ -77,7 +76,7 @@ public class MainLayoutController implements Initializable{
     private void loadScreen(String fxml) {
         new Thread(() -> {
             try {
-                FXMLLoader loader = new FXMLLoader(getClass().getResource("/com/ricardo/views/" + fxml + ".fxml"));
+                FXMLLoader loader = new FXMLLoader(getClass().getResource("/edu/grupo8/views/" + fxml + ".fxml"));
                 Parent content = loader.load();
 
                 Platform.runLater(() -> {
