+++ src/main/java/edu/grupo8/components/MenuHighlightShape.java
@@ -1,4 +1,4 @@
-package com.ricardo.components;
+package edu.grupo8.components;
 
 import javafx.scene.paint.Color;
 import javafx.scene.shape.ClosePath;
