+++ src/main/resources/edu/grupo8/MainLayout.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.text.Font?>
 <?import org.kordamp.ikonli.javafx.FontIcon?>
 
-<HBox prefHeight="400.0" prefWidth="640.0" styleClass="main-hbox" stylesheets="@styles/mainlayout_style.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.ricardo.MainLayoutController">
+<HBox prefHeight="400.0" prefWidth="640.0" styleClass="main-hbox" stylesheets="@styles/mainlayout_style.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.grupo8.MainLayoutController">
    <children>
       <AnchorPane fx:id="sidebarAnchorPane" maxWidth="190.0" prefHeight="380.0" prefWidth="190.0" styleClass="sidebar-menu-anchorpane">
          <children>
