+++ src/main/resources/edu/grupo8/fonts/Inter-VariableFont_opsz,wght.ttf
