+++ src/main/resources/edu/grupo8/icons/configicon.png
