+++ src/main/resources/edu/grupo8/icons/diagram.png
