+++ src/main/resources/edu/grupo8/icons/equipamentosicon.png
