+++ src/main/resources/edu/grupo8/icons/logo.png
