+++ src/main/resources/edu/grupo8/views/Equipamentos.fxml
